fix: make adjustments according to feedback
* reduce size of testing bottleneck list items
* add gradient on component's background
* change path from /bot -> /testing-bot
* sync other components with the new route
* provide redirection on the old route
* add missing key attribute in CMDLoginForm component
